Update MoviesController.php
In more complex queries the lack of specification causes a Integrity constraint violation: 1052 in your database. This is because the Where clause becomes ambiguous. Specifying 'Movie.'[$name] ensures that the constructed SQL statement is more specific so that in complex statements the ambiguity issue is addressed. (In my project, my equivalent of Movies was BTMHM)